Moved `Format` and `TextKind` out of `element` mod
