+++ src/element/writer.rs
@@ -7,9 +7,9 @@ use crate::result::IonResult;
 
 use crate::element::{Element, Value};
 use crate::ion_writer::IonWriter;
+pub use crate::Format::*;
 use crate::IonType;
-pub use Format::*;
-pub use TextKind::*;
+pub use crate::TextKind::*;
 
 /// Serializes [`Element`] instances into some kind of output sink.
 pub trait ElementWriter {
@@ -76,22 +76,6 @@ where
             }
         }
     }
-}
-
-/// Whether or not the text is pretty printed or serialized in a more compact representation.
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub enum TextKind {
-    Compact,
-    Lines,
-    Pretty,
-}
-
-/// Basic configuration options for [`ElementWriter`] instances.
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub enum Format {
-    Text(TextKind),
-    Binary,
-    // TODO a mode for Json(TextKind)
 }
 
 #[cfg(test)]
